chore(bolt-cli): request confirmation also on register validators command
