+++ bolt-cli/src/commands/operators.rs
@@ -1,5 +1,3 @@
-use std::io::Write;
-
 use alloy::{
     network::EthereumWallet,
     primitives::{utils::format_ether, Bytes},
@@ -13,7 +11,7 @@ use crate::{
     cli::{
         Chain, EigenLayerSubcommand, OperatorsCommand, OperatorsSubcommand, SymbioticSubcommand,
     },
-    common::bolt_manager::BoltManagerContract,
+    common::{bolt_manager::BoltManagerContract, request_confirmation},
     contracts::{
         bolt::{
             BoltEigenLayerMiddleware,
@@ -266,36 +264,6 @@ impl OperatorsCommand {
     }
 }
 
-/// Asks whether the user wants to proceed further. If not, the process is exited.
-fn request_confirmation() {
-    #[cfg(test)]
-    return;
-
-    loop {
-        info!("Do you want to continue? (yes/no): ");
-
-        print!("Answer: ");
-        std::io::stdout().flush().expect("Failed to flush");
-
-        let mut input = String::new();
-        std::io::stdin().read_line(&mut input).expect("Failed to read input");
-
-        let input = input.trim().to_lowercase();
-
-        match input.as_str() {
-            "yes" | "y" => {
-                return;
-            }
-            "no" | "n" => {
-                info!("Aborting");
-                std::process::exit(0);
-            }
-            _ => {
-                println!("Invalid input. Please type 'yes' or 'no'.");
-            }
-        }
-    }
-}
 #[cfg(test)]
 mod tests {
     use crate::{
